pyproject.toml
[project]
name = "websearch-agent"
version = "0.1.0"
description = "Sistema automatizado para verificação de licenciamento de softwares corporativos através de pesquisa web assistida por IA"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.7",
    "langchain-community>=0.0.20",
    "langchain-openai>=0.0.5",
    "langgraph>=0.0.20",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "office365-rest-python-client>=2.5.0",
    "openpyxl>=3.1.2",
    "pandas>=2.1.0",
    "python-dotenv>=1.0.0",
    "duckduckgo-search>=4.1.0",
]
